+++ apps/backend/src/libs/constants/constants.ts
@@ -1,1 +1,1 @@
-export { EMPTY_ARRAY_LENGTH, MAX_FILE_SIZE_IN_MB } from "@trackmates/shared";
+export { EMPTY_LENGTH, MAX_FILE_SIZE_IN_MB } from "@trackmates/shared";