+++ apps/backend/src/modules/friends/friend.repository.ts
@@ -1,12 +1,12 @@
 import { type Page } from "objection";
 
+import { EMPTY_LENGTH } from "~/libs/constants/constants.js";
 import { SortOrder } from "~/libs/enums/enums.js";
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
 import {
 	type PaginationResponseDto,
 	type Repository,
 } from "~/libs/types/types.js";
-import { EMPTY_ARRAY_LENGTH } from "~/libs/types/types.js";
 import { GroupEntity } from "~/modules/groups/group.entity.js";
 import { PermissionEntity } from "~/modules/permissions/permissions.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
@@ -213,7 +213,7 @@ class FriendRepository implements Repository<UserEntity> {
 			.where(`${DatabaseTableName.FRIENDS}.follower_id`, "=", currentUserId)
 			.where(`${DatabaseTableName.FRIENDS}.following_id`, "=", otherUserId);
 
-		return userFollowings.length > EMPTY_ARRAY_LENGTH;
+		return userFollowings.length > EMPTY_LENGTH;
 	}
 
 	public async getIsSubscribedByRequestId(
@@ -265,7 +265,13 @@ class FriendRepository implements Repository<UserEntity> {
 					.where(`${DatabaseTableName.FRIENDS}.follower_id`, "=", id)
 					.whereNotNull(`${DatabaseTableName.FRIENDS}.follower_id`),
 			)
-			.distinct()
+			.groupBy(
+				`${DatabaseTableName.USERS}.id`,
+				"userDetails.id",
+				"userDetails:avatar_file.id",
+				"groups.id",
+				"groups:permissions.id",
+			)
 			.withGraphJoined(
 				`[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
 			)