+++ apps/frontend/src/pages/chats/libs/helpers/get-formatted-chat-date/get-formatted-chat-date.helper.ts
@@ -18,13 +18,10 @@ const getFormattedChatDate = (dateString: string): string => {
 	}
 
 	if (checkIsThisYear(date)) {
-		return getFormattedDate(dateString, FormatDateType.MMMM_DO).toLowerCase();
+		return getFormattedDate(dateString, FormatDateType.MMMM_DO);
 	}
 
-	return getFormattedDate(
-		dateString,
-		FormatDateType.MMMM_DO_YYYY,
-	).toLowerCase();
+	return getFormattedDate(dateString, FormatDateType.MMMM_DO_YYYY);
 };
 
 export { getFormattedChatDate };